+++ packages/format-library/src/text-color/inline.js
@@ -110,7 +110,7 @@ function ColorPicker( { name, property, value, onChange } ) {
 	const colors = useSelect( ( select ) => {
 		const { getSettings } = select( blockEditorStore );
 		return get( getSettings(), [ 'colors' ], [] );
-	} );
+	}, [] );
 	const onColorChange = useCallback(
 		( color ) => {
 			onChange(